pyproject.toml
[tool.poetry]
name = "ml-core-template"
version = "0.1.0"
description = ""
authors = ["Karina Assini <karinaa@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "< 2.0.0"
pandas = "^2.1.3"
pydantic-yaml = "^1.2.0"
pytest-cov = "^4.1.0"
pyarrow = "^14.0.1"
pyyaml = "^6.0.1"
scikit-learn = "^1.3.2"
pre-commit = "^3.6.0"
lightgbm = "^4.2.0"
xgboost = "^2.0.3"
optuna = "^3.5.0"
shap = "^0.45.1"
matplotlib = "3.8.4"
azure-ai-ml = "1.18.0"
ipykernel = "6.29.5"
black="24.4.2"
mltable="1.6.1"
mlflow="2.15.0"
azureml-mlflow="1.53.0"
azureml-core="1.56.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
